Implement missing peripherals for Portal emulation
Partially implement the following peripherals required to emulate Portal:
- RNG
- PWR
- RTC
- I2C
- TSC
- GPIO INTERRUPT
- FLASH

I2C, TSC and GPIO INTERRUPT are accessible from the outside through
virtual serial ports.

The flash can be mapped to an MTD-like flash file on the host. The bank
switching logic is implemented and the SYSCFG_MEMRMP register correctly
configured.